test(postgres): isolate SSH tunnel live coverage
